+++ R/glyexp-package.R
@@ -3,10 +3,13 @@
 
 ## usethis namespace: start
 #' @import glyrepr
+#' @import methods
+#' @importFrom lifecycle deprecated
 #' @importFrom rlang :=
 #' @importFrom rlang .data
 #' @importFrom rlang %||%
 #' @importFrom tidyselect all_of
+#' @importClassesFrom SummarizedExperiment SummarizedExperiment
 ## usethis namespace: end
 NULL